backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-types==0.7.0 ; python_version >= "3.11"
anyio==4.8.0 ; python_version >= "3.11"
async-timeout==5.0.1 ; python_version >= "3.11" and python_version < "3.11.3"
certifi==2024.12.14 ; python_version >= "3.11"
charset-normalizer==3.4.1 ; python_version >= "3.11"
click==8.1.8 ; python_version >= "3.11"
colorama==0.4.6 ; platform_system == "Windows" and python_version >= "3.11" or sys_platform == "win32" and python_version >= "3.11"
coloredlogs==15.0.1 ; python_version >= "3.11"
dnspython==2.7.0 ; python_version >= "3.11"
email-validator==2.2.0 ; python_version >= "3.11"
fastapi-cli==0.0.7 ; python_version >= "3.11"
fastapi==0.115.6 ; python_version >= "3.11"
filelock==3.17.0 ; python_version >= "3.11"
fire==0.7.0 ; python_version >= "3.11"
flatbuffers==25.1.24 ; python_version >= "3.11"
fsspec==2025.2.0 ; python_version >= "3.11"
fugashi==1.4.0 ; python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11"
httpcore==1.0.7 ; python_version >= "3.11"
httptools==0.6.4 ; python_version >= "3.11"
httpx==0.28.1 ; python_version >= "3.11"
huggingface-hub==0.28.1 ; python_version >= "3.11"
humanfriendly==10.0 ; python_version >= "3.11"
idna==3.10 ; python_version >= "3.11"
jaconv==0.4.0 ; python_version >= "3.11"
jinja2==3.1.5 ; python_version >= "3.11"
libvalkey==4.0.1 ; python_version >= "3.11"
loguru==0.7.2 ; python_version >= "3.11"
manga-ocr==0.1.14 ; python_version >= "3.11"
markdown-it-py==3.0.0 ; python_version >= "3.11"
markupsafe==3.0.2 ; python_version >= "3.11"
mdurl==0.1.2 ; python_version >= "3.11"
mpmath==1.3.0 ; python_version >= "3.11"
networkx==3.4.2 ; python_version >= "3.11"
numpy==2.2.1 ; python_version >= "3.11"
onnxruntime==1.20.1 ; python_version >= "3.11"
opencv-python-headless==4.11.0.86 ; python_version >= "3.11"
packaging==24.2 ; python_version >= "3.11"
pillow==11.1.0 ; python_version >= "3.11"
protobuf==5.29.3 ; python_version >= "3.11"
pydantic-core==2.27.2 ; python_version >= "3.11"
pydantic==2.10.5 ; python_version >= "3.11"
pygments==2.19.1 ; python_version >= "3.11"
pyperclip==1.9.0 ; python_version >= "3.11"
pyreadline3==3.5.4 ; python_version >= "3.11" and sys_platform == "win32"
python-dotenv==1.0.1 ; python_version >= "3.11"
python-multipart==0.0.20 ; python_version >= "3.11"
pyyaml==6.0.2 ; python_version >= "3.11"
regex==2024.11.6 ; python_version >= "3.11"
requests==2.32.3 ; python_version >= "3.11"
rich-toolkit==0.12.0 ; python_version >= "3.11"
rich==13.9.4 ; python_version >= "3.11"
safetensors==0.5.2 ; python_version >= "3.11"
setuptools==75.8.0 ; python_version >= "3.12"
shellingham==1.5.4 ; python_version >= "3.11"
sniffio==1.3.1 ; python_version >= "3.11"
starlette==0.41.3 ; python_version >= "3.11"
sympy==1.13.1 ; python_version >= "3.11"
termcolor==2.5.0 ; python_version >= "3.11"
tokenizers==0.21.0 ; python_version >= "3.11"
torch==2.6.0+cpu ; python_version >= "3.11"
tqdm==4.67.1 ; python_version >= "3.11"
transformers==4.48.2 ; python_version >= "3.11"
typer==0.15.1 ; python_version >= "3.11"
typing-extensions==4.12.2 ; python_version >= "3.11"
unidic-lite==1.0.8 ; python_version >= "3.11"
urllib3==2.3.0 ; python_version >= "3.11"
uvicorn==0.34.0 ; python_version >= "3.11"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11"
valkey==6.0.2 ; python_version >= "3.11"
watchfiles==1.0.4 ; python_version >= "3.11"
websockets==14.1 ; python_version >= "3.11"
win32-setctime==1.2.0 ; python_version >= "3.11" and sys_platform == "win32"
